OpenDonationDialog moves from donations namespace to brave_rewards
